[mocks] Provide a message if the wrong double is used.
If an undefined method is stubbed on an instance double, the error generator
will check to see if it is defined on the class level and provide a
message to indicate that this may be what was meant. It also does the
converse for class doubles and instance methods.

https://github.com/rspec/rspec-mocks/issues/838

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/cd51da8ec07349f0e8257e36fa38ee2dc94605cf.