[core] Assign example group constant earlier.
This ensures it is set as soon as possible, before
shared contexts included via metadata are evaluated,
so that if there’s an error, the example group class
name will be included in it.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/5dc5088cfbb220762b1bdfa27392a64271de139c.