[core] Skip specs that are failing on AppVeyor on Ruby 2.1.
I have no idea why these pass on 1.9.3 but not on 2.1 on
AppVeyor.  It would be nice to get them to pass but
I don’t have access to a windows box and lack to the time
fix it now.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/361b6d5f0104665097c171525c859963de0dfa9a.