[core] Refactor: centralize exception presenter options.
We need to combine some of these cases (such
as when we get a `MultipleExpectationsNotMetError`
in a pending spec), and to do that we need to combine
the options, so having the options listed in the same
method is a stepping stone towards that.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/58925e5eeac77f07b7444681ff32faeeb4577c81.