[core] Make backtrace truncation much more robust.
This is necessary for the changes in rspec/rspec-support#210.

That adds an additional stack frame to the parent exception
(for the lambda that wraps `raise`) and our truncation wasn’t
working properly with it. Really, the old way just happened to
work. This is much more robust.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/59a30e8daef84e13966502420efd517597459d12.