[core] Always verify mocks at the end of each example.
...even if the example has already failed.

Previously, we did not have a good way to display
multiple failures for a single example, but we
do now, so it’s simpler and more consistent to
always verify, and to provide the user with multiple
exceptions if there are multiple.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/f989168c9dbbcd930af6e7cc61e08057194929a8.