[expectations] Work around travis build failing on REE.
The `include` cuke has a bunch of assertions that are failing
on REE since 1.8.7 does not guarantee hash ordering, so when
a hash is inspected, it could produce `{ :b => 2, :a => 1 }`
instead of `{ :a => 1, :b => 2 }`.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/8c403a07f150f9ecffaffbbecb95901c1d387bf0.