[core] Test the number of rounds to 'reduce' pathological case
If an expected failure depends on a large number of the passing examples, the
subset enumeration strategy can take a large number of rounds to minimise the
run.

Add a test to count rounds in the worst case, where a failure depends on all
of the preceding examples. With 8 candidates preceding a failure, a recursive
strategy could return in 15 rounds (1 + 2 + 4 + 8).

---
This commit was imported from https://github.com/rspec/rspec-core/commit/13f89e3a7a9baccca22c6eca1bd0890fc1c4ba39.