[support] Include delegators in inspection output
In current spec output, it’s not immediately obvious when an object
under test is actually wrapped in a delegator of some sort, which can
lead to some tricky-to-track-down failures on equality checks for two
seemingly-identical objects.

The fix proposed here is to toss objects into a special wrapper when we
prepare them for inspection. This wrapper recursively modifies the
object’s inspection string with delegator class information, which
should make it much more apparent when the object under test is wrapped
by a delegator of some sort.

---
This commit was imported from https://github.com/rspec/rspec-support/commit/6c3300a7cf24d952c3b456b078d3982da7db962c.