[support] Fix spec so it always passes on 1.8.7.
It failed on travis:

https://travis-ci.org/rspec/rspec-support/jobs/67086677

Hashes are unordered on 1.8.7 so the output order is non-deterministic.
For this spec it is not important that the hash has multiple entries,
so limiting it to one makes the output on all rubies deterministic.

---
This commit was imported from https://github.com/rspec/rspec-support/commit/0c8e070373cba79b91fb6d20e0f99d6763e48068.