[core] Update Rubocop to latest version.
This also updates the travis setup to use the latest bundler. Travis
currently defaults to Bundler version 1.7.4, however, this bundler
version does not support the `platform: :ruby_22` configuration setting.
So we need to tell travis to install a more recent bundler.

Also, `platform: :ruby_19` applies to **both** 1.9.2 and 1.9.3. But
Rubocop stopped supporting 1.9.2. Thus we need to check the
`RUBY_VERSION` in our `Gemfile` and not attempt to install Rubocop
specifically for 1.9.2.

Unfortunately, this bundler does not work with JRuby (2.0 mode), but
does work with JRuby 1.8 mode. So we need to update the config to ignore
JRuby 2.0. However, the current config incorrectly assumes we can do
that by checking the `JRUBY_OPTS` environment variable. Travis actually
sets this be default on _all_ systems:
http://docs.travis-ci.com/user/environment-variables/#Default-Environment-Variables

However, we can check the Ruby version we pass into the travis matrix by
using the `TRAVIS_RUBY_VERSION` and seeing if it equals "jruby".

---
This commit was imported from https://github.com/rspec/rspec-core/commit/6be271bc998005767d519c157b981460177e3aff.