[expectations] Be more selective about what exceptions we rescue.
Some exceptions should not be rescued. The exception
is the `raise_error` matcher — to allow people to
specify that a block of code raises a one of the
errors we normally don’t rescue we have to make it
rescue all errors.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/88e99d81b109a3c8787cb9d271f5964f24696309.