[expectations] Omit error when `nil` from warning in raise_error
Previously when the actual error was not present in the raise_error
matcher, the warning message generated would like this:
"Actual error raised was nil"

This commit removes the line above from the warning message, when the
cause of the error is `nil`.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/aadd2ddb8e76a6c48306dad0e67152b1431db424.