[core] Add `currently_executing_a_context_hook?` to ExampleGroup
This is useful because at the moment the only way to determine if an
example is currently executing a context hook is to look at it's inspect
string. Here, we add a simple boolean flag which is triggered in before
and after context hook execution.

This is needed by https://github.com/rspec/rspec-rails/pull/1501.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/f8e555dc4762df49029488ee94aaabbea98a5ec5.