[core] Fix DRb specs that were failing for me locally.
As the comment explains, I have had to configure my network
adapter in a non-standard way, and it causes
`IPSocket.getaddress(Socket.gethostname)` to raise an error
for me.

A couple of the DRb specs were failing as a result. The use of
`return` from `ensure` caused the error to be ignored (the fact
that `local_drb` was not set caused it to return false). This was
confusing, and using `return` from `ensure` is a horrible practice. I have refactored to avoid the issue, rescuing
errors explicitly.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/824fdd97d7ef439b3cd86823b18a39f9258d9765.