[expectations] Fix deprecation check for `should` with BasicObject.
The deprecation check uses `__method__` which is a Kernel method
that is not available on BasicObject. We can call it directly on
Kernel instead so that we avoid any `method_missing` chains that
may attempt to dispatch it to BasicObject.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/e91b893491cbbd96495c2fecc11bbd189c1964c3.