[core] Fix NoMethodError for NilClass for Java backtraces
On JRuby, the source of the error can come from Java code (JRuby internals).
In that case the line format is not guaranteed to follow the Ruby format, making the Regex fail.
In that case, let's handle that gracefully instead of raising an exception that hides the original cause.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/ad8751294d15abe39d955a79043eda9b2a65496f.