[core] Disable PerceivedComplexity cop for Parser#parse
This method is now occasionally (but not reliably) failing the
PerceivedComplexity cop. Since the complexity of this method is hard to
control and the CyclomaticComplexity rule is already disabled for it, I'm
turning off the cop for this method only.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/b5741ab4ae95bf5912de9e0ab7a74277aa6b05a3.