[core] Extract Parser::Callables to Invocations module
Previous work introduced a set of invocation callables that could be set in
the OptionParser to run instead of the default RSpec::Core::Runner. Testing
them via the the OptionParser seemed out-of-place, however, and made the test
stimulus needlessly complex.

This change extracts the invocations to a set of callable objects defined in
the RSpec::Core::Invocations namespace, allowing them to be tested separately
from the OptionParser.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/d52c96939306846f00be33649143dcda7863368e.