[core] Handle errors in :suite hooks.
Previously, we just allowed the error to propagate to
the user, which was a subpar experience for a few reasons:

- The error would cause RSpec to crash, leading to a long
  stack trace containing lots of extraneous info the user
  did not see.
- The error was not formatted nicely like other errors
  that happen while RSpec runs.
- If the error happened in an `after(:suite)` hook, the test
  suite had finished running all specs but the summary of the
  results would not get printed since RSpec had crashed.

Now, we handle errors in :suite hooks and format the output
the same way failures and errors are normally printed.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/77f8b01af5526f52c73771cc083109cf49aedc9c.