[core] Improve output from `rspec -v`.
Before this change, this confusing thing could happen:

```
$ gem install rspec -v 3.5.0 > /dev/null; rspec -v
3.5.1
```

The user installed RSpec 3.5.0 but `rspec -v` prints 3.5.1.
This happened because 3.5.0 of the rspec gem depends on 3.5.x
of all the sub-gems, and 3.5.1 is the latest rspec-core.

With this change, it now prints something like:

```
$ rspec -v
RSpec 3.5
  - rspec-core 3.5.1
  - rspec-expectations 3.5.0
  - rspec-mocks 3.5.0
  - rspec-support 3.5.0
```

---
This commit was imported from https://github.com/rspec/rspec-core/commit/e448575372ff1522dfb1c918547382eaec328dd7.