[core] Fix incorrect context and example descriptions
The context on line 62 had the same description as the one on line 1249 - "with config.tty? and !output.tty?" - but its implementation was the opposite - "with !config.tty? and output.tty?". Also, the example description was inverted - it said it "does not set color_enabled?", but it does. This commit corrects these descriptions to match the spec code.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/d4dcc57e9f2d8512b325c3e04fdc88f6a1811c5e.