[core] Specify color not enabled when option not passed
The spec previously did not specify the behavior of the color_enabled? flag when the color option was set to false. Because we will be significantly changing the behavior of color flags, specifying this existing behavior will help document any changes to it for backward-compatibility considerations.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/494ce68429d290238413a02f54c2216318eab7c4.