[core] Remove `example` from list of RSpec keywords.
While it is an alias of `it` and `specify`, it is
not often used, and often is used as a variable name
(e.g. when an `example` is yielded to a hook), so it
is better not to highlight it.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/74bbc1dd2e2833f5b11af4c44b30008698fb2f0e.