[core] Re-initializing the reporter when clearing examples.
This was prompted by rspec/rspec-core#2367 and the cumulative reporting out of
deprecations. Semantically, this could be weird! Should clearing
examples also clear the reporter? If the reporter is configured after
initialization through a means other than standard RSpec.configure, that
configuration would be lost. I don't know why you would do that though?

The original intent of clear_examples (from rspec/rspec-core#1700) seems to imply that
reseting the reporters was desirable, so I think this is OK.

It's weird that reporter belongs to configuration. It mixes state with
configuration. This means that we have to do a "partial" configuration
reset when we want to "reset except configuration", which suggests to me
that we'll probably have more issues here in the future.

Reported (among other things) in rspec/rspec-core#2367

---
This commit was imported from https://github.com/rspec/rspec-core/commit/1effb6ba64d23eb18b4a217a870bab3843619b81.