[mocks] Add a nice error message when calling stub_const with something other than a string (rspec/rspec-mocks#1148)
* Added a nice error message when calling  with something other than a string

* Moved unless statement to the top of the conditional check. Updated specs. Use === for type check instead of

Note that  may not work if the constant_name is a class that overrides that method, or maybe it doesnt respond to it at all. With ===, we know for sure that it will work for any type check

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/be8d5c92a200328addf5365d3421d2940eb711b5.