[core] Work around regressions introduced by lazy proc allocation.
This Ruby 2.5 feature is causing bugs with our hooks because
we depend upon `Hook#==` working properly for the same source
hook block. In Ruby 2.5 they introduced lazy proc allocation
which can result in two `Hook` instances which were previously
correctly considered to be equal to no longer be considered
equal. To work around this, we just need to invoke a method on
the proc before passing it along to other methods.

Note that there might be other bugs introduced by the Ruby 2.5
change, but this fixes the only test failures due to it, so this
is all we are changing for now.

Fixes rspec/rspec-core#2488.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/84670489bb4943a62e783bd65f96e4b55360b141.