[core] Allow users to set a formatter instance instead of just a class.
This is useful for when you want to initialize your formatter
with extra state. The simplest approach is for you to instantiate
the formatter yourself instead of depending on RSpec to do it and
trying to update the formatter with your desired state later.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/1488497c2f516912157aa641005d1af26fc23b82.