[expectations] Only call to_hash when actual object does not respond to include?
This fixes a regression introduced with commit a0c3958 that occurs
when an object has both an `include?` and a `to_hash` method. The
include matcher was delegating to the `include?` method on the Hash
returned from a call to `to_hash` instead of first checking if the
object implements `include?`.

A manual test was conducted to make sure that this change does not
regress PR rspec/rspec-expectations#1069. An instance of ActiveSupport::HashWithIndifferentAccess
was used in that test.

---
This commit was imported from https://github.com/rspec/rspec-expectations/commit/a8467cc7e6fd285ced741e218154b6bef0db222a.