[core] Clarify @tag (cuke style) functionality (rspec/rspec-core#2573)
With the previous documentation, it was unclear if the passed in tag
using the `@` (cucumber style tagging) would be ignored similar to doing
`~tag` (for `--tag @tag`), or if it would "ignore" the `@` when building
the configuration option (`@tag` changes to `:tag`).

Looking at the implementation in `lib/rspec/core/option_parser.rb`
confirms that it is the latter, so the wording has been updated to make
that more clear.

---
This commit was imported from https://github.com/rspec/rspec-core/commit/7288702e029bd22a8a3b8d98f3f0a22aef562253.