[support] Treat only symbol keys as keyword arguments
If you pass key-value pairs to a method only the keys that are Symbols
will be treated as keyword arguments. If there are optional arguments
the rest of the keys will be passed to the optional one.

Given a method foo(x = {}, y: 1):
* foo('a' => 1) => will work
* foo('a' => 1, :y => 2) => will work
* foo('a' => 1, :b => 2) => will raise ArgumentError unknown keyword: b

---
This commit was imported from https://github.com/rspec/rspec-support/commit/778f143da167ed4f3f8e8715543ddec153694aaa.