[mocks] Make `with` aware of Ruby 3 keywords
This change makes `receive(:foo).with(a: "A")` allow only `foo(a: "A")`
and deny `foo({ a: "A" })` in Ruby 3.
Also, `receive(:foo).with({ a: "A" })` allows only `foo({ a: "A" })` and
denies `foo(a: "A")`.

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/15153a6d56d0b40c1303e22202371fc69ee19ad7.