[mocks] Respect Ruby 3's conversion from keywords to a positional last Hash
This change allows a Hash argument expectation (`with({ a: 'a' })`) to
accept keyword arguments (`foo(a: 'a')`) both in Ruby 2 and in Ruby 3.

```
allow(foo).to receive(:bar).with({ a: 'a' })

foo.bar(a: 'a')
foo.bar({ a: 'a' })
```

This is because Ruby 3 also allows the automatic conversion from
keywords to a positional Hash. (But the conversion from a positional
last Hash to keywords is not allowed.)

```
def foo(opts = {})
end

foo(a: 'a')
foo({ a: 'a' })
```

---
This commit was imported from https://github.com/rspec/rspec-mocks/commit/20499b9784c61a3ceb357bd7054984938f4847ec.